requirements.txt
moviepy==1.0.3
requests>=2.31.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
numpy>=1.24.0
